+++ Changelog.md
@@ -1,3 +1,8 @@
+Simple Quests 1.3.1
+================
+- Add some more context to some stuff (API)
+- Add crafting task. Triggered when the player crafts an item
+
 Simple Quests 1.3.0
 ================
 - Add quest categories